backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pg8000==1.31.2
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.9
reportlab==4.2.5
openpyxl==3.1.5
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
